Updated
Use Case Description for Upload Audio Files from being worded in a such a way as to require parallel uploading to the point of view of operator and inputs and outputs across boundary.

Requester and Audio Analyzer System Interface Design to have Amalgamation as an output of use case Listen  and View.